feat: relocate documentation manager
